services/audio/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai-whisper>=20230918
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy<2.0
ffmpeg-python>=0.2.0